% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG.R
\name{payment_DRG}
\alias{payment_DRG}
\title{Payment DRG Function}
\usage{
payment_DRG(df, payment)
}
\arguments{
\item{df}{a data frame}

\item{payment}{can select: Average Covered Charges, Average Total Payments, or Average Medicare Payments. Must be in " " when using.}
}
\value{
A box plot with all DRG codes on the x axis and payment parameter plotted on y axis
}
\description{
Payment DRG Function
}
\examples{
library(data.table)
df.test <- data.table::fread('https://data.cms.gov/api/views/97k6-zzx3/rows.csv?accessType=DOWNLOAD')
payment_DRG(df.test, "Average Covered Charges")

}
