% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG.R
\name{stat_DRG}
\alias{stat_DRG}
\title{Statistic DRG Function}
\usage{
stat_DRG(df, stat, DRG_code)
}
\arguments{
\item{df}{Data frame}

\item{stat}{Select statistic of interest: Mean, median, or standard deviation. Must be in " " when using.}

\item{DRG_code}{Select DRG code of interest. Must be in " " when using.}
}
\value{
The mean, median, or sd of average medicare payments for DRG code of interest
}
\description{
Statistic DRG Function
}
\examples{
library(data.table)
df.test <- data.table::fread('https://data.cms.gov/api/views/97k6-zzx3/rows.csv?accessType=DOWNLOAD')
stat_DRG(df.test, "mean", "065")

}
